Updates for consistent conventions
rng issue addressed.  When we want unfrozen noise, we use the rng to select a random seed to pass to the coneMosaic compute method, with appropriate saving and restoring.  This gives the behavior desired at this level without needing any chances to coneMosiac.  rng state is also saved/restored in cases where frozen noise is requested.

Also adopted and implemented convention that just one instance is returned in the no noise form of the call.